% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{annotate.paths}
\alias{annotate.paths}
\title{Annotates functions to pathways}
\usage{
annotate.paths(metaginfo, dbannot)
}
\arguments{
\item{metaginfo}{Pathways object}

\item{dbannot}{Either a string indicating which precomputed annotation to
use ("uniprot" for Uniprot Keywords or "GO" for Gene Ontology terms), or
a dataframe with the annotation of the genes to the functions. First
column are gene symbols, second column the functions.}
}
\value{
Object of annotations from pathways to functions
}
\description{
Annotates functions from a database to each pathway
}
\examples{

}
