% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brca_design}
\alias{brca_design}
\title{BRCA experimental design}
\format{Dataframe with 1 column and 40 rows, including the experimental
design of the 40 samples from the BRCA-US project from TCGA. Field
\code{group} is the type of sample, either "Tumor" or "Normal".}
\source{
\url{https://cancergenome.nih.gov/}
}
\usage{
data(brca_design)
}
\value{
Dataframe with 1 column and 40 rows, including the experimental
design of the 40 samples from the BRCA-US project from TCGA. Field
\code{group} is the type of sample, either "Tumor" or "Normal".
}
\description{
Experimental design of the gene expression matrix \code{brca_data} with
40 samples taken from the BRCA-US project from The Cancer Genome Atlas
(TCGA). 20 samples are "Tumor" samples and 20 samples are "Normal" samples.
}
\keyword{datasets}
