% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comp}
\alias{comp}
\title{Wilcoxon comparison of pathways object}
\format{Table with 1868 rows and 5 columns}
\usage{
data(comp)
}
\value{
Pathway comparison result
}
\description{
Comparison object returned by \code{hipathia::do_wilcoxon} function, after
calling
\code{comp <- do_wilcoxon(path_vals, sample_group, g1 = "Tumor", g2 =
"Normal")}
\code{path_names <- get_path_names(pathways, rownames(comp))}
\code{comp <- cbind(path_names, comp)}
}
\keyword{datasets}
