% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DAdata}
\alias{DAdata}
\title{Wilcoxon and limma comparison object for nodes, pathways and functional
annotations}
\format{
List object with 4 entries:
Nodes includes a matrix with 6826 rows and 8 columns
Paths includes a matrix with 1876 rows and 13 columns
Uni.terms includes a matrix with 142 rows and 6 columns
GO.terms includes a matrix with 1654 rows and 6 columns
}
\usage{
data(DAdata)
}
\value{
List of tibbles with the comparison results
}
\description{
Comparison object returned by \code{hipathia::DAcomp} function, after
calling
\code{DAdata <- DAcomp(hidata, "group", g1 = "Tumor", g2 = "Normal")}
}
\keyword{datasets}
