% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel.R
\name{define_colors}
\alias{define_colors}
\title{Color palettes to be used in plots.}
\usage{
define_colors(colors, no.col = NULL)
}
\arguments{
\item{colors}{String with the color scheme or vector of colors to be used.
Available predefined options include: \code{hipathia}, \code{classic},
\code{soft}, \code{okee}, \code{hiro}, \code{new}, \code{vg}, \code{orchid}.}

\item{no.col}{String with the color given to non-significant nodes, if not
given in parameter \code{colors}.}
}
\value{
Plot and list of tables including top \code{n} altered features for
nodes, paths and functions if present.
}
\description{
Color palettes to be used in plots.
}
\examples{
define_colors("hiro")

}
