% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{do.pca}
\alias{do.pca}
\title{Performs a Principal Components Analysis}
\usage{
do.pca(data, cor = FALSE)
}
\arguments{
\item{data}{Matrix of values to be analyzed. Samples must be represented
in the columns.}

\item{cor}{A logical value indicating whether the calculation should use
the correlation matrix or the covariance matrix. (The correlation matrix
can only be used if there are no constant variables.)}
}
\value{
\code{do.pca} returns a list with class \code{princomp}.
}
\description{
Performs a Principal Components Analysis
}
\examples{
data(path_vals)
pca.model <- do.pca(path_vals[1:ncol(path_vals),])

}
