% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{do.wilcoxon}
\alias{do.wilcoxon}
\title{Apply Wilcoxon test}
\usage{
do.wilcoxon(sel.vals, group.value, g1, g2, paired = FALSE, adjust = TRUE)
}
\arguments{
\item{sel.vals}{Matrix of values. Columns represent samples.}

\item{group.value}{Vector with the class to which each sample belongs.
Samples must be ordered as in \code{sel.vals}}

\item{g1}{String, label of the first group to be compared}

\item{g2}{String, label of the second group to be compared}

\item{paired}{Boolean, whether the samples to be compared are paired.
If TRUE, function \code{wilcoxsign_test} from package \code{coin} is
used. If FALSE, function \code{wilcox.test} from package \code{stats}
is used.}

\item{adjust}{Boolean, whether to adjust the p.value with
Benjamini-Hochberg FDR method}
}
\value{
Dataframe with the result of the comparison
}
\description{
Performs a Wilcoxon test for the values in \code{sel.vals} comparing
conditions \code{g1} and \code{g2}
}
\examples{
data(path_vals)
data(brca_design)
sample.group <- brca_design[colnames(path_vals),"group"]
comp <- do.wilcoxon(path_vals, sample.group, g1 = "Tumor", g2 = "Normal")

}
