% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp_data}
\alias{exp_data}
\title{Normalized BRCA gene expression dataset}
\format{Matrix with 40 columns and 3184 rows. Row names are Entrez IDs
and column names are the  TCGA identifyers of the samples.}
\usage{
data(exp_data)
}
\value{
Matrix with 40 columns and 3184 rows. Row names are Entrez IDs
and column names are the  TCGA identifyers of the samples.
}
\description{
Experimental design matrix once expression matrix \code{brca_data} has been
translated to Entrez geens with \code{translate_matrix} and normalized using
\code{normalize_data}.
}
\details{
To create the data, the following functions have been called:
\code{trans_data <- translate_matrix(brca_data, "hsa")}
\code{exp_data <- normalize_data(trans_data)}
}
\keyword{datasets}
