% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.pathway.functions}
\alias{get.pathway.functions}
\title{Returns functions related to a pathway}
\usage{
get.pathway.functions(pathigraph, dbannot, entrez2hgnc, use.last.nodes = TRUE,
  unique = TRUE)
}
\arguments{
\item{pathigraph}{Pathway object}

\item{dbannot}{Dataframe with the annotation of the genes to the functions.
First column are gene symbols, second column the functions.}
}
\value{
List of annotations from pathways to functions
}
\description{
Returns functions related to a pathway
}
