% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.pathways.list}
\alias{get.pathways.list}
\title{Lists the IDs of the pathways in a pathways object}
\usage{
get.pathways.list(metaginfo)
}
\arguments{
\item{metaginfo}{Pathways object}
}
\value{
List of the pathway IDs included in the pathways object
}
\description{
Lists the IDs of the pathways included in the pathways object
\code{metaginfo}
}
\examples{
pathways <- load.pathways(species = "hsa", pathways.list = c("hsa03320",
"hsa04012"))
pathways.list <- get.pathways.list(pathways)

}
