% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get.pathways.summary}
\alias{get.pathways.summary}
\title{Compute pathway summary}
\usage{
get.pathways.summary(comp, metaginfo, conf = 0.05)
}
\arguments{
\item{comp}{Comparison data frame as returned by the \code{do.wilcoxon}
function.}

\item{metaginfo}{Pathways object}

\item{conf}{Level of significance of the comparison for the adjusted
p-value. Default is 0.05.}
}
\value{
Table with the summarized information for each of the pathways.
Rows are the analized pathways. Columns are:
* \code{num.total.paths} Number of total subpathways in which each pathway
is decomposed.
* \code{num.significant.paths} Number of significant subpathways in the
provided comparison.
* \code{percent.significant.paths} Percentage of significant subpathways
from the total number of subpathways in a pathway.
* \code{num.up.paths} Number of significant up-regulated subpathways in the
provided comparison.
* \code{percent.up.paths} Percentage of significant up-regulated subpathways
from the total number of subpathways in a pathway.
* \code{num.down.paths} Number of significant down-regulated subpathways in
the provided comparison.
* \code{percent.down.paths} Percentage of significant down-regulated
subpathways from the total number of subpathways in a pathway.
}
\description{
Computes a summary of the results, summarizing the number and proportion
of up- and down-regulated subpathways in each pathway.
}
\examples{
data(comp)
pathways <- load.pathways(species = "hsa")
get.pathways.summary(comp, pathways)

}
