% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{get_file}
\alias{get_file}
\title{Loads a RData object and downloads it from Zenodo if necessary}
\usage{
get_file(file)
}
\arguments{
\item{file}{File name of the object to load

#@examples
#get_file("xref_hsa_v3.rda")
#get_file("meta_graph_info_hsa_v3.rda")}
}
\value{
Object in the file
}
\description{
Loads a RData object and downloads it from Zenodo if necessary
}
