% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_highest_sig_ancestor}
\alias{get_highest_sig_ancestor}
\title{Get highest common GO ancestor of GO annotations}
\usage{
get_highest_sig_ancestor(
  go_terms,
  go_comp,
  metaginfo,
  unique = TRUE,
  pval = 0.05
)
}
\arguments{
\item{go_terms}{GO terms for which the highest common ancestors are
to be looked for.}

\item{go_comp}{Wilcoxon comparison of the matrix of GO values as returned
by \code{do_wilcoxon}.}

\item{metaginfo}{Pathways object}

\item{unique}{Boolean, whether to return only one highest significant GO
ancestor or all of them. By default, TRUE.}

\item{pval}{P-value cut-off. Default values is set to 0.05.}
}
\value{
highest common ancestors

#@export
}
\description{
Get highest common GO ancestor of GO annotations
}
