% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pathways_list}
\alias{get_pathways_list}
\title{Lists the IDs of the pathways in a pathways object}
\usage{
get_pathways_list(metaginfo)
}
\arguments{
\item{metaginfo}{Pathways object}
}
\value{
List of the pathway IDs included in the pathways object
}
\description{
Lists the IDs of the pathways included in the pathways object
\code{metaginfo}
}
\examples{
pathways <- load_pathways(species = "hsa", pathways_list = c("hsa03320",
"hsa04012"))
pathways_list <- get_pathways_list(pathways)

}
