% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{go_vals}
\alias{go_vals}
\title{Gene Ontology matrix of the BRCA gene expression dataset}
\format{Matrix with 40 columns and 1654 rows. Row names are Gene Ontology
terms and column names are the TCGA identifyers of the samples.}
\usage{
data(go_vals)
}
\value{
Matrix with 40 columns and 1654 rows. Row names are Gene Ontology
terms and column names are the TCGA identifyers of the samples.
}
\description{
Matrix of Gene Ontology terms activation values for the BRCA dataset.
This matrix is computed from the Results object returned by the
\code{hipathia} function by means of the \code{quantify_terms} function.
}
\details{
\code{go_vals <- quantify_terms(results, pathways, "GO")}
}
\keyword{datasets}
