% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hidata}
\alias{hidata}
\title{Results object}
\format{
MultiAssayExperiment object of 4 listed experiments, with the
activity values of nodes, paths and functional annotations for each sample:
Nodes includes a matrix with 6826 rows
Paths includes a matrix with 1876 rows
Uni.terms includes a matrix with 142 rows
GO.terms includes a matrix with 1654 rows
}
\usage{
data(hidata)
}
\value{
Object of results, including nodes, pathways and functional
information.
}
\description{
Results object returned by \code{hipathia::hipathia} function, after calling
\code{hidata <- hipathia(brca, pathways, verbose=TRUE, uni.terms = TRUE,
GO.terms = TRUE)}
}
\keyword{datasets}
