% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{igraphs.upgrade}
\alias{igraphs.upgrade}
\title{Upgrade igraphs to current version}
\usage{
igraphs.upgrade(metaginfo)
}
\arguments{
\item{metaginfo}{Pathways object}
}
\value{
The pathways object with the upgraded igraph objects
}
\description{
Upgrades the \code{igraph} objects in metaginfo object to the corresponding
version of the \code{igraph} package.
}
