% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{is_accepted_species}
\alias{is_accepted_species}
\title{Checks whether a species is accepted}
\usage{
is_accepted_species(species)
}
\arguments{
\item{species}{Species of the samples.

#@examples
#is_accepted_species("hsa")
#is_accepted_species("fca")}
}
\value{
Boolean, whether \code{species} is accepted or not.
}
\description{
Checks whether a species is accepted
}
