% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.annots}
\alias{load.annots}
\title{Loads functional annotations}
\usage{
load.annots(db, species)
}
\arguments{
\item{db}{Database to be used. Either "GO" or "uniprot".}

\item{species}{Species of the samples.}
}
\value{
Functional annotations from HGNC to the selected database.
}
\description{
Loads functional annotations from HGNC to the selected database.
}
\examples{
load.annots("GO", "hsa")

}
