% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.entrez.hgnc}
\alias{load.entrez.hgnc}
\title{Loads table of translation from HGNC to Entrez}
\usage{
load.entrez.hgnc(species)
}
\arguments{
\item{species}{Species of the samples.

#@examples
#load.entrez.hgnc("hsa")}
}
\value{
Table of translation from HGNC to Entrez
}
\description{
Loads table of translation from HGNC to Entrez
}
