% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.gobp.frame}
\alias{load.gobp.frame}
\title{Loads GO graph information}
\usage{
load.gobp.frame()
}
\value{
GO graph information
}
\description{
Loads GO graph information
}
\examples{
load.gobp.frame()

}
