% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.mgi}
\alias{load.mgi}
\title{Loads object with graph information}
\usage{
load.mgi(species)
}
\arguments{
\item{species}{Species of the samples.

#@examples
#load.mgi("hsa")}
}
\value{
Graph information object
}
\description{
Loads object with graph information
}
