% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.unibp}
\alias{load.unibp}
\title{Loads Uniprot annotations}
\usage{
load.unibp(species)
}
\arguments{
\item{species}{Species of the samples.}
}
\value{
Uniprot annotations
}
\description{
Loads Uniprot annotations
}
