% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.xref}
\alias{load.xref}
\title{Loads table of references}
\usage{
load.xref(species)
}
\arguments{
\item{species}{Species of the samples.}
}
\value{
Table of references
}
\description{
Loads table of references
}
\examples{
load.xref("hsa")

}
