% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_annofuns}
\alias{load_annofuns}
\title{Loads annotations object}
\usage{
load_annofuns(db, species)
}
\arguments{
\item{db}{Database to be used. Either "GO" or "uniprot".}

\item{species}{Species of the samples.

#@examples
#load_annofuns("GO", "hsa")
#load_annofuns("uniprot", "hsa")}
}
\value{
Annotations object
}
\description{
Loads annotations object
}
