% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_entrez_hgnc}
\alias{load_entrez_hgnc}
\title{Loads table of translation from HGNC to Entrez}
\usage{
load_entrez_hgnc(species)
}
\arguments{
\item{species}{Species of the samples.

#@examples
#load_entrez_hgnc("hsa")}
}
\value{
Table of translation from HGNC to Entrez
}
\description{
Loads table of translation from HGNC to Entrez
}
