% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_mgi}
\alias{load_mgi}
\title{Loads object with graph information}
\usage{
load_mgi(species)
}
\arguments{
\item{species}{Species of the samples.

#@examples
#load_mgi("hsa")}
}
\value{
Graph information object
}
\description{
Loads object with graph information
}
