% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_pseudo_mgi}
\alias{load_pseudo_mgi}
\title{Loads object with pseudo graph information}
\usage{
load_pseudo_mgi(species, group_by)
}
\arguments{
\item{species}{Species of the samples.}

\item{group_by}{How to group the subpathways to be visualized. By default
they are grouped by the pathway to which they belong. Available groupings
include "uniprot", to group subpathways by their annotated Uniprot functions,
"GO", to group subpathways by their annotated GO terms, and "genes", to group
subpathways by the genes they include.

#@examples
#load_pseudo_mgi("hsa", "uniprot")}
}
\value{
Pseudo graph information object
}
\description{
Loads object with pseudo graph information
}
