% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_xref}
\alias{load_xref}
\title{Loads table of references}
\usage{
load_xref(species)
}
\arguments{
\item{species}{Species of the samples.

#@examples
#load_xref("hsa")}
}
\value{
Table of references
}
\description{
Loads table of references
}
