% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{normalize.data}
\alias{normalize.data}
\title{Normalize expression data to be used in \code{hipathia}}
\usage{
normalize.data(exp.data, by.quantiles = FALSE, by.gene = FALSE,
  percentil = FALSE, truncation.percentil = NULL)
}
\arguments{
\item{exp.data}{Matrix of gene expression.}

\item{by.quantiles}{Boolean, whether to normalize the data by quantiles.
Default is FALSE.}

\item{by.gene}{Boolean, whether to transform the rank of each row of the
matrix to [0,1]. Default is FALSE.}

\item{percentil}{Boolean, whether to take as value the percentil of each
sample in the corresponding distribution.}

\item{truncation.percentil}{Real number p in [0,1]. When provided, values
beyond percentil p are truncated to the value of percentil p, and values
beyond 1-p are truncated to percentil 1-p. By default no truncation
is performed.}
}
\value{
Matrix of gene expression whose values are in [0,1].
}
\description{
Transforms the rank of the matrix of gene expression to [0,1] in order
to be processed by \code{hipathia}. The transformation may be performed
in two different ways. If \code{percentil = FALSE}, the transformation
is a re-scaling of the rank of the matrix. If \code{percentil = TRUE},
the transformation is performed assigning to each cell its percentil in
the corresponding distribution. This option is recommended for
distributions with very long tails.
}
\details{
This transformation may be applied either to the whole matrix
(by setting \code{by.gene = FALSE}), which we strongly recommend, or to
each of the rows (by setting \code{by.gene = TRUE}), allowing each gene
to have its own scale.

A previous quantiles normalization may be applied by setting
\code{by.quantiles = TRUE}. This is recommended for noisy data.

For distributions with extreme outlayer values, a percentil \code{p}
may be given to the parameter \code{truncation.percentil}. When provided,
values beyond percentil p are truncated to the value of percentil p, and
values beyond 1-p are truncated to percentil 1-p. This step is performed
 before any other tranformation. By default no truncation is performed.
}
\examples{
data("brca_data")
trans.data <- translate.matrix(brca_data, "hsa")
exp.data <- normalize.data(trans.data)
exp.data <- normalize.data(trans.data, by.quantiles = TRUE,
truncation.percentil=0.95)

}
