% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize.paths}
\alias{normalize.paths}
\title{Normalize the pathway matrix by rows}
\usage{
normalize.paths(path.vals, metaginfo)
}
\arguments{
\item{path.vals}{Matrix of the pathway values}

\item{metaginfo}{Pathways object}
}
\value{
Matrix of normalized pathway values
}
\description{
Due to the nature of the Hipathia method, the length of a pathway may
influence its signal rank. In order to compare signal values among
subpathways, we strongly recommend to normalize the matrix with this
normalization.
}
\details{
This function removes the bias caused by the length of the subpathways
by dividing by the value obtained from running the method with a basal
value of 0.5 at each node.
}
\examples{
data(path_vals)
pathways <- load.pathways(species = "hsa", pathways.list = c("hsa03320",
"hsa04012"))
path.normalized <- normalize.paths(path_vals, pathways)

}
