% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{path_vals}
\alias{path_vals}
\title{Pathways matrix of the BRCA gene expression dataset}
\format{Matrix with 40 columns and 1868 rows. Row names are Pathway IDs
and column names are the TCGA identifyers of the samples.}
\usage{
data(path_vals)
}
\value{
Matrix with 40 columns and 1868 rows. Row names are Pathway IDs
and column names are the TCGA identifyers of the samples.
}
\description{
Matrix of pathway activation values for the BRCA dataset. This matrix is
extracted from the Results object returned by the \code{hipathia} function
by means of the \code{get_paths_matrix} function.
}
\details{
\code{path_vals <- get_paths_matrix(results)}
}
\keyword{datasets}
