% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathways}
\alias{pathways}
\title{Pathways object including pathways has03320 and hsa04012.}
\format{
Pathways object
}
\usage{
data(pathways)
}
\value{
Pathways object including pathways has03320 and hsa04012.
}
\description{
Pathways object returned by \code{hipathia::load_pathways} function, after
calling
\code{pathways <- load_pathways(species = "hsa",
pathways_list = c("hsa03320", "hsa04012"))}
}
\keyword{datasets}
