% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{results}
\alias{results}
\title{Results object}
\format{Object of results, including pathways information.}
\usage{
data(results)
}
\value{
Object of results, including pathways information.
}
\description{
Results object returned by \code{hipathia::hipathia} function, after calling
\code{results <- hipathia(exp_data, pathways, verbose=TRUE)}
}
\keyword{datasets}
