% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save.results}
\alias{save.results}
\title{Save results to folder}
\usage{
save.results(results, comp, metaginfo, output.folder)
}
\arguments{
\item{results}{Results object as returned by the \code{hipathia} function.}

\item{comp}{Comparison as returned by the \code{do.wilcoxon} function.}

\item{metaginfo}{Pathways object}

\item{output.folder}{Absolute path to the folder in which results will be
saved. If this folder does not exist, it will be created.
However, the parent folder must exist.}
}
\value{
Creates a folder in disk in which all the information to browse the
pathway results is stored.
}
\description{
Saves results to a folder. In particular, it saves the matrix of subpathway
values, a table with the results of the provided comparison,
the accuracy of the results and the .SIF and attributes of the pathways.
}
\examples{
data(results)
data(comp)
pathways <- load.pathways(species = "hsa")
save.results(results, comp, pathways, "output.results")

}
