% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{top_pathways}
\alias{top_pathways}
\title{Computes pathway significance}
\usage{
top_pathways(comp)
}
\arguments{
\item{comp}{Comparison data frame as returned by the \code{do_wilcoxon}
function.}
}
\value{
Table with the names of the pathways and their p-value for the Fisher test
comparing the proportion of significant subpaths vs. 0.
}
\description{
Performs a test for each pathway checking if the number of significant
paths is significant, compared to not having any of the paths as significant.
}
\examples{
data(comp)
top_pathways(comp)

}
