% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{translate.matrix}
\alias{translate.matrix}
\title{Translation of the rownames IDs to Entrez IDs.}
\usage{
translate.matrix(exp, species, verbose = TRUE)
}
\arguments{
\item{exp}{Matrix of gene expression.}

\item{species}{Species of the samples.}

\item{verbose}{Boolean, whether to show details about the results of the
execution.}
}
\value{
Matrix of gene expression with Entrez IDs as rownames.
}
\description{
Translates the IDs in the rownames of a matrix to Entrez IDs.
For accepted IDs to be transformed see the DOCUMENTATION.
}
\examples{
data("brca_data")
trans.data <- translate.matrix(brca_data, "hsa")

}
