% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uni_vals}
\alias{uni_vals}
\title{Uniprot matrix of the BRCA gene expression dataset}
\format{Matrix with 40 columns and 142 rows. Row names are Uniprot functions
and column names are the TCGA identifyers of the samples.}
\usage{
data(uni_vals)
}
\value{
Matrix with 40 columns and 142 rows. Row names are Uniprot functions
and column names are the TCGA identifyers of the samples.
}
\description{
Matrix of Uniprot functions activation values for the BRCA dataset.
This matrix is computed from the Results object returned by the
\code{hipathia} function by means of the \code{quantify_terms} function.
}
\details{
\code{uni_vals <- quantify_terms(results, pathways, "uniprot")}
}
\keyword{datasets}
