% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{visualize.report}
\alias{visualize.report}
\title{Visualize a HiPathia report}
\usage{
visualize.report(output.folder, port = 4000)
}
\arguments{
\item{output.folder}{Folder in which results to visualize are stored}

\item{port}{Port to use}
}
\value{
The instructions to visualize a HiPathia report in a web browser
}
\description{
Visualize a HiPathia report
}
\examples{
data(results)
data(brca_design)
data(path_vals)
pathways <- load.pathways(species = "hsa", pathways.list = c("hsa03320",
"hsa04012"))
sample.group <- brca_design[colnames(path_vals),"group"]
colors.de <- node.color.per.de(results, pathways,
sample.group, "Tumor", "Normal")
create.report(results, comp, pathways, "~/save_results/",
node.colors = colors.de)
visualize.report("~/save_results/")
visualize.report("~/save_results/", port=5000)
\dontshow{servr::daemon_stop()}

}
