% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{create.report}
\alias{create.report}
\title{Create visualization HTML}
\usage{
create.report(results, comp, metaginfo, output.folder, node.colors = NULL,
  conf = 0.05, verbose = FALSE)
}
\arguments{
\item{results}{Results object as given by the \code{hipathia} function}

\item{comp}{Comparison object as given by the \code{do.wilcoxon} function}

\item{metaginfo}{Pathways object as returned by the \code{load.pathways}
function}

\item{output.folder}{Absolute path to the folder in which results will be
saved. If this folder does not exist, it will be created.
However, the parent folder must exist.}

\item{node.colors}{List of colors with which to paint the nodes of the
pathways, as returned by the
\code{node.color.per.differential.expression} function. Default is white.}

\item{conf}{Level of significance. By default 0.05.}

\item{verbose}{Boolean, whether to show details about the results of the
execution}
}
\value{
Saves the results and creates a report to visualize them through
a server in the specified \code{output.folder}.
}
\description{
Saves the results of a Wilcoxon comparison for the Hipathia pathway values
into a folder, and creates a HTML from which to visualize the results on
top of the pathways. The results are stored into the specified folder.
If this folder does not exist, it will be created. The parent folder must
exist.
}
\examples{
data(results)
data(comp)
data(brca_design)
data(path_vals)
pathways <- load.pathways(species = "hsa")
create.report(results, comp, pathways, "~/save_results/")

sample_group <- brca_design[colnames(path_vals),"group"]
colors.de <- node.color.per.differential.expression(results, pathways,
sample_group, "Tumor", "Normal")
create.report(results, comp, pathways, "~/save_results/",
node.colors = colors.de)

}
