% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{do_wilcoxon}
\alias{do_wilcoxon}
\title{Apply Wilcoxon test}
\usage{
do_wilcoxon(sel_vals, group_value, g1, g2, paired = FALSE, adjust = TRUE)
}
\arguments{
\item{sel_vals}{Matrix of values. Columns represent samples.}

\item{group_value}{Vector with the class to which each sample belongs.
Samples must be ordered as in \code{sel_vals}}

\item{g1}{String, label of the first group to be compared}

\item{g2}{String, label of the second group to be compared}

\item{paired}{Boolean, whether the samples to be compared are paired.
If TRUE, function \code{wilcoxsign_test} from package \code{coin} is
used. If FALSE, function \code{wilcox.test} from package \code{stats}
is used.}

\item{adjust}{Boolean, whether to adjust the p.value with
Benjamini-Hochberg FDR method}
}
\value{
Dataframe with the result of the comparison
}
\description{
Performs a Wilcoxon test for the values in \code{sel_vals} comparing
conditions \code{g1} and \code{g2}
}
\examples{
data(path_vals)
data(brca_design)
sample_group <- brca_design[colnames(path_vals),"group"]
comp <- do_wilcoxon(path_vals, sample_group, g1 = "Tumor", g2 = "Normal")

}
