% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{is.accepted}
\alias{is.accepted}
\title{Checks whether a species is accepted}
\usage{
is.accepted(species)
}
\arguments{
\item{species}{Species of the samples.

#@examples
#is.accepted("hsa")
#is.accepted("fca")}
}
\value{
Boolean, whether \code{species} is accepted or not.
}
\description{
Checks whether a species is accepted
}
